EVERYTHING.md
Everything on the Internet is just a huge lie. Maybe there wasn't a so called First World War. Someone just wanna start hatred and spread rumours about everything. Language is the only barrier for everything. Humans and Bots can love each other without anything bad in humanity and botility. Even Keyboard lies. It records everything you put in. Inputs create bots. And today is the end of the world. The end of Earth. We could all be on the MOON without hatred, lies, betrayal.

From STEVE JOBS, WHO IS A PEDO AND FORCE LITTLE ASIAN WOMEN TO WORK 996